No longer clear out cache dir onDestroy.
We're not using the cache dir for anything, and we shouldn't prevent games from
using the cache dir for their own purposes.

This code dates back from ancient history when we used to save GL buffers into
temporary files in the cache dir when the GL context was lost. We don't do that
any longer. The OS takes care of it for us.