Simplify API for glGetActive{Attrib|Uniform}.
This used to only be implemented on LWJGL2+3 & Android, now it's only
implemented on LWJGL2+3 because to do it on Android would require that we
depend on API 17 (4.2) which is the first API for which Google no longer
published the android.jar file via Maven.

Since clearly no one was using this broken, partially implemented API, I'm not
going to go through all the trouble to try to get it working on Android.