Undo last commit, add what is really needed.
Custom users of RoboStorage need to pass in a way for RoboStorage to log
things. They don't (necessarily) need to subclass it and touch the log shim.

I removed the constructor with the NOOP logger since a custom user would
ideally make their own choice about whether or not to ignore log messages.