Fix rendering issue in fillRoundRect/strokeRoundRect
The current implementation translates the canvas, draws using a rect at
(0, 0), and then translates it back. This doesn't play nice with gradient
fills.

The translate calls were introduced to fix a bug when the rect supplied
to drawRoundRect had non-zero x or y [1]. However, the problem was actually
related to the coords passed to the rect (the two last parameters should
be the right and bottom coords, not the width and height).

So fix the fix by passing the right coords to the rect and removing the
translate calls.

 [1]: ab72fb60845312a2cb2815874f1ad995d89d90f0