ScaledResource: fix support for scaling factors < 1
The computePath method was not adjusting the resource path for scaling
factors < 1. This is necessary for Android ldpi devices (yes, there are
some of these still around) where the scaling factor is 0.75.

Also, when generating the scale suffix for scaling factors < 1, prepend
a "0" to differentiate from suffixes corresponding to factors > 1.