Fix text rendering for non-integer font sizes
Set the LINEAR_TEXT_FLAG to fix text rendering when using non-integer
font sizes. Without this, the extent of text runs for non-integer
font sizes seems to be calculated based on integer sizes, and the
letter spacing seems to be wrong.

After this change, the rendering of text in Android is much closer
to the results when using PlayN on robovm/iOS.